Added Swagger documentation
- Moved main endpoints to `routes.js` from app.js
- Moved helper functions from app.js
- Added packages for `swagger-jsdoc`
- Added endpoint documentation
- Added `/api-docs` endpoint
