+++ src/__tests__/integration-tests/article1036/article1036-test.js
@@ -11,7 +11,7 @@ import { prettify } from '../../testHelpers';
 import article1036 from './article1036';
 import image2357 from './image2357';
 
-import { fetchAndTransformArticle } from '../../../app';
+import { fetchAndTransformArticle } from '../../../api/helpers';
 
 test('app/fetchAndTransformArticle 1036', async () => {
   nock('https://test.api.ndla.no')
