+++ src/__tests__/integration-tests/article417/article417-test.js
@@ -12,7 +12,7 @@ import { prettify } from '../../testHelpers';
 import article417 from './article417';
 import articleResource413 from './articleResource413';
 import articleResource414 from './articleResource414';
-import { fetchAndTransformArticle } from '../../../app';
+import { fetchAndTransformArticle } from '../../../api/helpers';
 import log from '../../../utils/logger';
 
 const resources = {
