+++ src/api/helpers/index.js
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) 2016-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import { fetchAndTransformArticle } from './helpers';
+
+export { fetchAndTransformArticle };
