+++ src/api/swagger/swagger.yaml
@@ -1,0 +1,14 @@
+parameters:
+  authorizationHeader:
+    in: header
+    name: authorization
+    description: Authorization token.
+    required: true
+    type: string
+
+securityDefinitions:
+    oauth2:
+      type: oauth2
+      description:
+      flow: implicit
+      authorizationUrl: https://ndla.eu.auth0.com/authorize
